requirements.txt
awscli==1.3.3
boto==2.27.0
paramiko==1.13.0
PyYAML==3.10
